examples: refactor the tensorflow-imagenet-resnet50 example
Refactors the tensorflow-imagenet-resnet50 demo to use the new, common approach designed for the
examples folder. This includes the following changes:

- Moved any new custom task plugins from tensorflow-imagenet-resnet50 folder to the root of the
  examples folder.
- Added the new builtin plugin extract_tarfile_in_unique_subdir to support extracting tarfiles into
  a unique subfolder in the current working directory.
- Moved the MLproject file into the src folder.
- Refactored the task plugins and entrypoint scripts to remove all direct usages of Prefect tasks.
- Pruned ununsed task plugins and entrypoint scripts from the example.
- Updated the documentation in examples/tensorflow-imagenet-resnet50/README.md with
  instructions on how the user should prepare their environment for running the examples.
- Modified the demo Jupyter notebook for tensorflow-imagenet-resnet50 to use the new scripts.
- Removed any files from the tensorflow-imagenet-resnet50 folder that were no longer being used
  after implementing the above changes.
