+++ examples/task-plugins/dioptra_custom/custom_fgm_plugins/__init__.py
