+++ examples/task-plugins/dioptra_custom/custom_fgm_plugins/attacks_fgm.py
@@ -27,20 +27,13 @@ import structlog
 from structlog.stdlib import BoundLogger
 
 from dioptra import pyplugs
-from dioptra.sdk.exceptions import (
-    ARTDependencyError,
-    TensorflowDependencyError,
-)
+from dioptra.sdk.exceptions import ARTDependencyError, TensorflowDependencyError
 from dioptra.sdk.utilities.decorators import require_package
 
 LOGGER: BoundLogger = structlog.stdlib.get_logger()
 
 try:
-    from art.defences.preprocessor import (
-        GaussianAugmentation,
-        JpegCompression,
-        SpatialSmoothing,
-    )
+    from art.attacks.evasion import FastGradientMethod
     from art.estimators.classification import KerasClassifier
 
 except ImportError:  # pragma: nocover
@@ -59,116 +52,39 @@ except ImportError:  # pragma: nocover
         package="tensorflow",
     )
 
-DEFENSE_LIST = {
-    "spatial_smoothing": SpatialSmoothing,
-    "jpeg_compression": JpegCompression,
-    "gaussian_augmentation": GaussianAugmentation,
-}
-
-
-def get_optimizer(optimizer, learning_rate):
-    optimizer_collection = {
-        "rmsprop": RMSprop(learning_rate),
-        "adam": Adam(learning_rate),
-        "adagrad": Adagrad(learning_rate),
-        "sgd": SGD(learning_rate),
-    }
-
-    return optimizer_collection.get(optimizer)
-
-
-# Load model from registry and apply imagenet_preprocessing if needed.
-def wrap_keras_classifier(model, clip_values, imagenet_preprocessing):
-    keras_model = load_model_in_registry(model=model)
-    if imagenet_preprocessing:
-        mean_b = 103.939
-        mean_g = 116.779
-        mean_r = 123.680
-        return KerasClassifier(
-            model=keras_model,
-            clip_values=clip_values,
-            preprocessing=(np.array([mean_b, mean_g, mean_r]), np.array([1.0, 1.0, 1.0])),
-        )
-    else:
-        return KerasClassifier(model=keras_model, clip_values=clip_values)
-
-
-def init_defense(clip_values, def_type, **kwargs):
-
-    defense = DEFENSE_LIST[def_type](
-        clip_values=clip_values,
-        **kwargs,
-    )
-
-    return defense
-
-
-def save_def_batch(def_batch, def_data_dir, y, clean_filenames, class_names_list):
-    for batch_image_num, def_image in enumerate(def_batch):
-        out_label = class_names_list[y[batch_image_num]]
-        def_image_path = (
-            def_data_dir
-            / f"{out_label}"
-            / f"def_{clean_filenames[batch_image_num].name}"
-        )
-
-        if not def_image_path.parent.exists():
-            def_image_path.parent.mkdir(parents=True)
-
-        save_img(path=str(def_image_path), x=def_image)
-
-
-def evaluate_distance_metrics(
-    clean_filenames, distance_metrics_, clean_batch, def_batch
-):
-    LOGGER.debug("evaluate image perturbations using distance metrics")
-    distance_metrics_["image"].extend([x.name for x in clean_filenames])
-    distance_metrics_["label"].extend([x.parent for x in clean_filenames])
-    for metric_name, metric in DISTANCE_METRICS:
-        distance_metrics_[metric_name].extend(metric(clean_batch, def_batch))
-
-
-def log_distance_metrics(distance_metrics_):
-    distance_metrics_ = distance_metrics_.copy()
-    del distance_metrics_["image"]
-    del distance_metrics_["label"]
-    for metric_name, metric_values_list in distance_metrics_.items():
-        metric_values = np.array(metric_values_list)
-        mlflow.log_metric(key=f"{metric_name}_mean", value=metric_values.mean())
-        mlflow.log_metric(key=f"{metric_name}_median", value=np.median(metric_values))
-        mlflow.log_metric(key=f"{metric_name}_stdev", value=metric_values.std())
-        mlflow.log_metric(
-            key=f"{metric_name}_iqr", value=scipy.stats.iqr(metric_values)
-        )
-        mlflow.log_metric(key=f"{metric_name}_min", value=metric_values.min())
-        mlflow.log_metric(key=f"{metric_name}_max", value=metric_values.max())
-        LOGGER.info("logged distance-based metric", metric_name=metric_name)
-
 
 @pyplugs.register
 @require_package("art", exc_type=ARTDependencyError)
 @require_package("tensorflow", exc_type=TensorflowDependencyError)
-def create_defended_dataset(
+def create_adversarial_fgm_dataset(
     data_dir: str,
-    def_data_dir: Union[str, Path],
+    adv_data_dir: Union[str, Path],
+    keras_classifier: KerasClassifier,
     image_size: Tuple[int, int, int],
     distance_metrics_list: Optional[List[Tuple[str, Callable[..., np.ndarray]]]] = None,
+    rescale: float = 1.0 / 255,
     batch_size: int = 32,
     label_mode: str = "categorical",
-    def_type: str = "spatial_smoothing",
-    **kwargs,
+    eps: float = 0.3,
+    eps_step: float = 0.1,
+    minimal: float = 0,
+    norm: float = np.inf,
+    target_index: int = -1,
+    targeted: bool = False,
 ) -> pd.DataFrame:
     distance_metrics_list = distance_metrics_list or []
     color_mode: str = "rgb" if image_size[2] == 3 else "grayscale"
-    rescale: float = 1.0 if image_size[2] == 3 else 1.0 / 255
-    clip_values: Tuple[float, float] = (0, 255) if image_size[2] == 3 else (0, 1.0)
     target_size: Tuple[int, int] = image_size[:2]
-    def_data_dir = Path(def_data_dir)
+    adv_data_dir = Path(adv_data_dir)
 
-    defense = init_defense(
-        clip_values=clip_values,
-        def_type=def_type,
-        **kwargs,
+    attack = _init_fgm(
+        keras_classifier=keras_classifier,
+        batch_size=batch_size,
+        eps=eps,
+        eps_step=eps_step,
+        minimal=minimal,
+        norm=norm,
+        targeted=targeted,
     )
 
     data_generator: ImageDataGenerator = ImageDataGenerator(rescale=rescale)
@@ -191,8 +107,8 @@ def create_defended_dataset(
         distance_metrics_[metric_name] = []
 
     LOGGER.info(
-        "Generate defended images",
-        defense=def_type,
+        "Generate adversarial images",
+        attack="fgm",
         num_batches=num_images // batch_size,
     )
 
@@ -205,49 +121,57 @@ def create_defended_dataset(
         ]
 
         LOGGER.info(
-            "Generate defended image batch",
-            defense=def_type,
+            "Generate adversarial image batch",
+            attack="fgm",
             batch_num=batch_num,
         )
 
         y_int = np.argmax(y, axis=1)
-        adv_batch_defend, _ = defense(x)
+        if target_index >= 0:
+            y_one_hot = np.zeros(n_classes)
+            y_one_hot[target_index] = 1.0
+            y_target = np.tile(y_one_hot, (x.shape[0], 1))
 
-        _save_def_batch(
-            adv_batch_defend, def_data_dir, y_int, clean_filenames, class_names_list
+            adv_batch = attack.generate(x=x, y=y_target)
+        else:
+            adv_batch = attack.generate(x=x)
+
+        _save_adv_batch(
+            adv_batch, adv_data_dir, y_int, clean_filenames, class_names_list
         )
 
         _evaluate_distance_metrics(
             clean_filenames=clean_filenames,
             distance_metrics_=distance_metrics_,
             clean_batch=x,
-            adv_batch=adv_batch_defend,
+            adv_batch=adv_batch,
             distance_metrics_list=distance_metrics_list,
         )
 
-    LOGGER.info("Defended image generation complete", defense=def_type)
+    LOGGER.info("Adversarial image generation complete", attack="fgm")
     _log_distance_metrics(distance_metrics_)
 
     return pd.DataFrame(distance_metrics_)
 
 
-def init_defense(clip_values, def_type, **kwargs):
-    defense = DEFENSE_LIST[def_type](
-        clip_values=clip_values,
-        **kwargs,
+def _init_fgm(
+    keras_classifier: KerasClassifier, batch_size: int, **kwargs
+) -> FastGradientMethod:
+    attack: FastGradientMethod = FastGradientMethod(
+        estimator=keras_classifier, batch_size=batch_size, **kwargs
     )
-    return defense
+    return attack
 
 
-def _save_def_batch(
-    adv_batch, def_data_dir, y, clean_filenames, class_names_list
+def _save_adv_batch(
+    adv_batch, adv_data_dir, y, clean_filenames, class_names_list
 ) -> None:
     for batch_image_num, adv_image in enumerate(adv_batch):
         out_label = class_names_list[y[batch_image_num]]
         adv_image_path = (
-            def_data_dir
+            adv_data_dir
             / f"{out_label}"
-            / f"def_{clean_filenames[batch_image_num].name}"
+            / f"adv_{clean_filenames[batch_image_num].name}"
         )
 
         if not adv_image_path.parent.exists():
