+++ examples/task-plugins/dioptra_custom/custom_fgm_plugins/data_tensorflow.py
@@ -16,33 +16,23 @@
 # https://creativecommons.org/licenses/by/4.0/legalcode
 from __future__ import annotations
 
-import mlflow
-import structlog
-import numpy as np
+from typing import Optional, Tuple
 
+import structlog
 from structlog.stdlib import BoundLogger
+from tensorflow.keras.applications.imagenet_utils import preprocess_input
 
 from dioptra import pyplugs
-from dioptra.sdk.exceptions import (
-    ARTDependencyError,
-    TensorflowDependencyError,
-)
+from dioptra.sdk.exceptions import TensorflowDependencyError
 from dioptra.sdk.utilities.decorators import require_package
 
 LOGGER: BoundLogger = structlog.stdlib.get_logger()
 
 try:
-    from art.estimators.classification import KerasClassifier
-
-except ImportError:  # pragma: nocover
-    LOGGER.warn(
-        "Unable to import one or more optional packages, functionality may be reduced",
-        package="art",
+    from tensorflow.keras.preprocessing.image import (
+        DirectoryIterator,
+        ImageDataGenerator,
     )
-
-
-try:
-    from tensorflow.keras.models import Sequential
 
 except ImportError:  # pragma: nocover
     LOGGER.warn(
@@ -51,35 +41,52 @@ except ImportError:  # pragma: nocover
     )
 
 
+# Adding imagenet preprocessing option for pre-trained Imagenet estimators.
 @pyplugs.register
-@require_package("art", exc_type=ARTDependencyError)
 @require_package("tensorflow", exc_type=TensorflowDependencyError)
-def load_wrapped_tensorflow_keras_classifier(
-    name: str,
-    version: int,
-    clip_values: Tuple = None,
+def create_image_dataset(
+    data_dir: str,
+    subset: Optional[str],
+    image_size: Tuple[int, int, int],
+    seed: int,
+    validation_split: Optional[float] = 0.2,
+    batch_size: int = 32,
+    label_mode: str = "categorical",
     imagenet_preprocessing: bool = False,
-) -> KerasClassifier:
+    rescale: float = 1.0,
+) -> DirectoryIterator:
+    color_mode: str = "rgb" if image_size[2] == 3 else "grayscale"
+    target_size: Tuple[int, int] = image_size[:2]
 
-    uri = f"models:/{name}/{version}"
-    LOGGER.info("Load Keras classifier from model registry", uri=uri)
-    keras_classifier = mlflow.keras.load_model(uri)
     if imagenet_preprocessing:
-        mean_b = 103.939
-        mean_g = 116.779
-        mean_r = 123.680
-
-        wrapped_keras_classifier = KerasClassifier(
-            model=keras_classifier,
-            clip_values=clip_values,
-            preprocessing=(np.array([mean_b, mean_g, mean_r]), np.array([1.0, 1.0, 1.0])),
+        data_generator: ImageDataGenerator = ImageDataGenerator(
+            rescale=rescale,
+            validation_split=validation_split,
+            preprocessing_function=preprocess_input,
         )
     else:
-        wrapped_keras_classifier = KerasClassifier(
-            model=keras_classifier, clip_values=clip_values
+        data_generator: ImageDataGenerator = ImageDataGenerator(
+            rescale=rescale,
+            validation_split=validation_split,
         )
-    LOGGER.info(
-        "Wrap Keras classifier for compatibility with Adversarial Robustness Toolbox"
+
+    return data_generator.flow_from_directory(
+        directory=data_dir,
+        target_size=target_size,
+        color_mode=color_mode,
+        class_mode=label_mode,
+        batch_size=batch_size,
+        seed=seed,
+        subset=subset,
     )
 
-    return wrapped_keras_classifier
+
+@pyplugs.register
+@require_package("tensorflow", exc_type=TensorflowDependencyError)
+def get_n_classes_from_directory_iterator(ds: DirectoryIterator) -> int:
+    return len(ds.class_indices)
+
+
+@pyplugs.register
+def rescale(imagenet_preprocessing: bool) -> float:
+    return 1.0 if imagenet_preprocessing else 1.0 / 255
