--- examples/tensorflow-imagenet-pixel-threshold/MLproject
@@ -1,50 +1,0 @@
-# This Software (Dioptra) is being made available as a public service by the
-# National Institute of Standards and Technology (NIST), an Agency of the United
-# States Department of Commerce. This software was developed in part by employees of
-# NIST and in part by NIST contractors. Copyright in portions of this software that
-# were developed by NIST contractors has been licensed or assigned to NIST. Pursuant
-# to Title 17 United States Code Section 105, works of NIST employees are not
-# subject to copyright protection in the United States. However, NIST may hold
-# international copyright in software created by its employees and domestic
-# copyright (or licensing rights) in portions of software that were assigned or
-# licensed to NIST. To the extent that NIST holds copyright in this software, it is
-# being made available under the Creative Commons Attribution 4.0 International
-# license (CC BY 4.0). The disclaimers of the CC BY 4.0 license apply to all parts
-# of the software developed or licensed by NIST.
-#
-# ACCESS THE FULL CC BY 4.0 LICENSE HERE:
-# https://creativecommons.org/licenses/by/4.0/legalcode
-name: imagenet-pixel-threshold
-
-entry_points:
-  pt:
-    parameters:
-      data_dir: {type: path, default: "/nfs/data/ImageNet-Kaggle-2017/images/ILSVRC/Data/CLS-LOC"}
-      model: {type: string, default: "keras-model-imagenet-resnet50/1"}
-      model_architecture: {type: string, default: "alex_net"}
-      batch_size: {type: float, default: 32}
-      th: {type: float, default: 1}
-      es: {type: float, default: 1}
-    command: >
-      python src/pt.py
-      --data-dir {data_dir}
-      --model {model}
-      --model-architecture {model_architecture}
-      --batch-size 25
-      --th 10
-      --es 0
-
-  init_model:
-    parameters:
-      data_dir: {type: path, default: "/nfs/data/ImageNet-Kaggle-2017/images/ILSVRC/Data/CLS-LOC/val-sorted-5000"}
-      model_tag: {type: string, default: "default_pretrained"}
-      model_architecture: {type: string, default: "resnet50"}
-      batch_size: {type: float, default: 10}
-      seed: {type: float, default: -1}
-    command: >
-      python src/init_model.py
-      --data-dir {data_dir}
-      --model-tag {model_tag}
-      --model-architecture {model_architecture}
-      --batch-size {batch_size}
-      --seed {seed}
