--- examples/tensorflow-imagenet-pixel-threshold/environment.yml
@@ -1,98 +1,0 @@
-# This Software (Dioptra) is being made available as a public service by the
-# National Institute of Standards and Technology (NIST), an Agency of the United
-# States Department of Commerce. This software was developed in part by employees of
-# NIST and in part by NIST contractors. Copyright in portions of this software that
-# were developed by NIST contractors has been licensed or assigned to NIST. Pursuant
-# to Title 17 United States Code Section 105, works of NIST employees are not
-# subject to copyright protection in the United States. However, NIST may hold
-# international copyright in software created by its employees and domestic
-# copyright (or licensing rights) in portions of software that were assigned or
-# licensed to NIST. To the extent that NIST holds copyright in this software, it is
-# being made available under the Creative Commons Attribution 4.0 International
-# license (CC BY 4.0). The disclaimers of the CC BY 4.0 license apply to all parts
-# of the software developed or licensed by NIST.
-#
-# ACCESS THE FULL CC BY 4.0 LICENSE HERE:
-# https://creativecommons.org/licenses/by/4.0/legalcode
-name: imagenet-pixel-threshold
-channels:
-  - defaults
-  - conda-forge
-dependencies:
-  - autopep8
-  - bottleneck
-  - click>=7.1.0
-  - cloudpickle
-  - cython
-  - entrypoints>=0.3
-  - flake8>=3.8.0
-  - flake8-bugbear
-  - h5py
-  - ipywidgets
-  - ipython
-  - isort>=5.6.0
-  - jedi=0.17.2 # Downgrade fixes IPython tab-completion error
-  - jinja2
-  - joblib
-  - jupyter
-  - jupyterlab
-  - markupsafe
-  - matplotlib
-  - mypy
-  - nodejs
-  - nomkl
-  - numexpr
-  - numpy>=1.19.0
-  - pandas>=1.1.0
-  - parso=0.7.* # Downgrade fixes IPython tab-completion error
-  - pillow
-  - pip
-  - protobuf
-  - pycodestyle
-  - pydocstyle
-  - redis-py
-  - pytables
-  - python=3.8.*
-  - pytz
-  - pyyaml=5.3.*
-  - requests=2.24.*
-  - scipy=1.4.1
-  - scikit-learn=0.23.*
-  - setuptools
-  - wheel
-  - pip:
-      - -e ../../.
-      - adversarial-robustness-toolbox==1.5.1
-      - alembic<=1.4.1
-      - black>=20.8b1
-      - boto3>=1.16.0
-      - databricks-cli
-      - docformatter
-      - Flask>=1.1.0
-      - flask-accepts>=0.17.0
-      - Flask-Injector>=0.12.0
-      - Flask-JWT-Extended>3.24.0
-      - Flask-Migrate>=2.5.0
-      - flask-restx>=0.2.0
-      - Flask-SQLAlchemy>=2.4.0
-      - Flask-WTF>=0.14.0
-      - gorilla
-      - injector>=0.18.0
-      - markdown
-      - marshmallow>=3.9.0,<4
-      - mlflow==1.12.1
-      - mnist
-      - multimethod>=1.5
-      - passlib
-      - prefect
-      - prometheus-flask-exporter
-      - pyarrow
-      - python-json-logger
-      - rich
-      - rq>=1.5.0
-      - sqlalchemy-stubs
-      - sqlalchemy==1.3.13
-      - structlog>=20.2.0
-      - tabulate
-      - tensorflow==2.1.0
-      - typing-extensions>=3.7.4.0
