+++ examples/tensorflow-imagenet-resnet50/src/fgm.py
@@ -17,15 +17,13 @@
 # https://creativecommons.org/licenses/by/4.0/legalcode
 
 import os
-import tarfile
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, List
 
 import click
 import mlflow
 import numpy as np
 import structlog
-from mlflow.tracking import MlflowClient
 from prefect import Flow, Parameter
 from prefect.utilities.logging import get_logger as get_prefect_logger
 from structlog.stdlib import BoundLogger
@@ -87,16 +85,26 @@ def _coerce_int_to_bool(ctx, param, value):
     help="Dimensions for the input images",
 )
 @click.option(
-    "--def-tar-name",
+    "--adv-tar-name",
     type=click.STRING,
-    default="guassian_augmentation_dataset.tar.gz",
-    help="Name to give to tarfile artifact containing preprocessed  images",
+    default="testing_adversarial_fgm.tar.gz",
+    help="Name to give to tarfile artifact containing fgm images",
 )
 @click.option(
-    "--def-data-dir",
+    "--adv-data-dir",
     type=click.STRING,
     default="adv_testing",
-    help="Directory for saving preprocessed images",
+    help="Directory for saving fgm images",
+)
+@click.option(
+    "--model-name",
+    type=click.STRING,
+    help="Name of model to load from registry",
+)
+@click.option(
+    "--model-version",
+    type=click.STRING,
+    help="Version of model to load from registry",
 )
 @click.option(
     "--batch-size",
@@ -105,58 +113,43 @@ def _coerce_int_to_bool(ctx, param, value):
     default=32,
 )
 @click.option(
-    "--gaussian-augmentation-perform-data-augmentation",
-    type=click.BOOL,
-    help="If set to true, original samples will be kept alongside noisy samples.",
-    default=False,
-)
-@click.option(
-    "--gaussian-augmentation-ratio",
+    "--eps",
     type=click.FLOAT,
-    help="If data augmentation is set to true, specifies fraction of new samples (1=double dataset size).",
-    default=1,
+    help="FGM attack step size (input variation)",
+    default=0.3,
 )
 @click.option(
-    "--gaussian-augmentation-sigma",
+    "--eps-step",
     type=click.FLOAT,
-    help="Standard deviation of Gaussian noise to be added.",
-    default=1,
+    help="FGM attack step size of input variation for minimal perturbation computation",
+    default=0.1,
 )
 @click.option(
-    "--gaussian-augmentation-apply-fit",
+    "--minimal",
+    type=click.Choice(["0", "1"]),
+    callback=_coerce_int_to_bool,
+    help="If 1, compute the minimal perturbation using eps_step for the step size and "
+    "eps for the maximum perturbation.",
+    default="0",
+)
+@click.option(
+    "--norm",
+    type=click.Choice(["inf", "1", "2"]),
+    default="inf",
+    callback=_map_norm,
+    help="FGM attack norm of adversarial perturbation",
+)
+@click.option(
+    "--imagenet-preprocessing",
     type=click.BOOL,
-    help="Defense applied on images used for training.",
+    help="If true, initializes model with Imagenet image preprocessing settings.",
     default=False,
 )
 @click.option(
-    "--gaussian-augmentation-apply-predict",
-    type=click.BOOL,
-    help="Defense applied on images used for testing.",
-    default=True,
-)
-@click.option(
-    "--load-dataset-from-mlruns",
-    type=click.BOOL,
-    help="If set to true, instead loads the test dataset from a previous mlrun.",
-    default=False,
-)
-@click.option(
-    "--dataset-run-id",
-    type=click.STRING,
-    help="MLFlow Run ID of an updated dataset.",
-    default="",
-)
-@click.option(
-    "--dataset-tar-name",
-    type=click.STRING,
-    help="Name of dataset tarfile.",
-    default="adversarial_poison.tar.gz",
-)
-@click.option(
-    "--dataset-name",
-    type=click.STRING,
-    help="Name of dataset directory.",
-    default="adv_poison_data",
+    "--target-index",
+    type=click.INT,
+    help="Class index for targeted attack. If set to a negative value, the fgm attack will be untargeted.",
+    default="-1",
 )
 @click.option(
     "--seed",
@@ -164,67 +157,64 @@ def _coerce_int_to_bool(ctx, param, value):
     help="Set the entry point rng seed",
     default=-1,
 )
-def guassian_augmentation(
+def fgm_attack(
     data_dir,
     image_size,
-    def_tar_name,
-    def_data_dir,
+    adv_tar_name,
+    adv_data_dir,
+    model_name,
+    model_version,
     batch_size,
-    gaussian_augmentation_perform_data_augmentation,
-    gaussian_augmentation_ratio,
-    gaussian_augmentation_sigma,
-    gaussian_augmentation_apply_fit,
-    gaussian_augmentation_apply_predict,
-    load_dataset_from_mlruns,
-    dataset_run_id,
-    dataset_tar_name,
-    dataset_name,
+    eps,
+    eps_step,
+    minimal,
+    norm,
+    imagenet_preprocessing,
+    target_index,
     seed,
 ):
-
+    targeted = False
+    if target_index >= 0:
+        targeted = True
     LOGGER.info(
         "Execute MLFlow entry point",
-        entry_point="guassian_augmentation",
+        entry_point="fgm",
         data_dir=data_dir,
         image_size=image_size,
-        def_tar_name=def_tar_name,
-        def_data_dir=def_data_dir,
+        adv_tar_name=adv_tar_name,
+        adv_data_dir=adv_data_dir,
+        model_name=model_name,
+        model_version=model_version,
         batch_size=batch_size,
-        gaussian_augmentation_perform_data_augmentation=gaussian_augmentation_perform_data_augmentation,
-        gaussian_augmentation_ratio=gaussian_augmentation_ratio,
-        gaussian_augmentation_sigma=gaussian_augmentation_sigma,
-        gaussian_augmentation_apply_fit=gaussian_augmentation_apply_fit,
-        gaussian_augmentation_apply_predict=gaussian_augmentation_apply_predict,
-        dataset_run_id=dataset_run_id,
-        dataset_tar_name=dataset_tar_name,
-        dataset_name=dataset_name,
+        eps=eps,
+        eps_step=eps_step,
+        minimal=minimal,
+        targeted=targeted,
+        target_index=target_index,
+        imagenet_preprocessing=imagenet_preprocessing,
+        norm=norm,
         seed=seed,
     )
 
-    if load_dataset_from_mlruns:
-        data_dir = Path.cwd() / "dataset" / dataset_name
-        data_tar_name = dataset_tar_name
-        data_tar_path = download_image_archive(
-            run_id=dataset_run_id, archive_path=data_tar_name
-        )
-        with tarfile.open(data_tar_path, "r:gz") as f:
-            f.extractall(path=(Path.cwd() / "dataset"))
-
     with mlflow.start_run() as active_run:  # noqa: F841
-        flow: Flow = init_guassian_augmentation_flow()
+        flow: Flow = init_fgm_flow()
         state = flow.run(
             parameters=dict(
                 testing_dir=Path(data_dir),
                 image_size=image_size,
-                def_tar_name=def_tar_name,
-                def_data_dir=(Path.cwd() / def_data_dir).resolve(),
+                adv_tar_name=adv_tar_name,
+                adv_data_dir=(Path.cwd() / adv_data_dir).resolve(),
                 distance_metrics_filename="distance_metrics.csv",
+                model_name=model_name,
+                model_version=model_version,
                 batch_size=batch_size,
-                gaussian_augmentation_perform_data_augmentation=gaussian_augmentation_perform_data_augmentation,
-                gaussian_augmentation_ratio=gaussian_augmentation_ratio,
-                gaussian_augmentation_sigma=gaussian_augmentation_sigma,
-                gaussian_augmentation_apply_fit=gaussian_augmentation_apply_fit,
-                gaussian_augmentation_apply_predict=gaussian_augmentation_apply_predict,
+                eps=eps,
+                eps_step=eps_step,
+                minimal=minimal,
+                norm=norm,
+                targeted=targeted,
+                target_index=target_index,
+                imagenet_preprocessing=imagenet_preprocessing,
                 seed=seed,
             )
         )
@@ -232,51 +222,48 @@ def guassian_augmentation(
     return state
 
 
-# Update data dir path if user is applying defense over image artifacts.
-def download_image_archive(
-    run_id: str, archive_path: str, destination_path: Optional[str] = None
-) -> str:
-    client: MlflowClient = MlflowClient()
-    image_archive_path: str = client.download_artifacts(
-        run_id=run_id, path=archive_path, dst_path=destination_path
-    )
-    LOGGER.info(
-        "Image archive downloaded",
-        run_id=run_id,
-        storage_path=archive_path,
-        dst_path=image_archive_path,
-    )
-    return image_archive_path
-
-
-def init_guassian_augmentation_flow() -> Flow:
+def init_fgm_flow() -> Flow:
     with Flow("Fast Gradient Method") as flow:
         (
             testing_dir,
             image_size,
-            def_tar_name,
-            def_data_dir,
+            adv_tar_name,
+            adv_data_dir,
             distance_metrics_filename,
+            model_name,
+            model_version,
             batch_size,
-            gaussian_augmentation_perform_data_augmentation,
-            gaussian_augmentation_ratio,
-            gaussian_augmentation_sigma,
-            gaussian_augmentation_apply_fit,
-            gaussian_augmentation_apply_predict,
+            eps,
+            eps_step,
+            minimal,
+            norm,
+            targeted,
+            target_index,
+            imagenet_preprocessing,
             seed,
         ) = (
             Parameter("testing_dir"),
             Parameter("image_size"),
-            Parameter("def_tar_name"),
-            Parameter("def_data_dir"),
+            Parameter("adv_tar_name"),
+            Parameter("adv_data_dir"),
             Parameter("distance_metrics_filename"),
+            Parameter("model_name"),
+            Parameter("model_version"),
             Parameter("batch_size"),
-            Parameter("gaussian_augmentation_perform_data_augmentation"),
-            Parameter("gaussian_augmentation_ratio"),
-            Parameter("gaussian_augmentation_sigma"),
-            Parameter("gaussian_augmentation_apply_fit"),
-            Parameter("gaussian_augmentation_apply_predict"),
+            Parameter("eps"),
+            Parameter("eps_step"),
+            Parameter("minimal"),
+            Parameter("norm"),
+            Parameter("targeted"),
+            Parameter("target_index"),
+            Parameter("imagenet_preprocessing"),
             Parameter("seed"),
+        )
+        rescale = pyplugs.call_task(
+            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.custom_fgm_plugins",
+            "data_tensorflow",
+            "rescale",
+            imagenet_preprocessing=imagenet_preprocessing,
         )
         seed, rng = pyplugs.call_task(
             f"{_PLUGINS_IMPORT_PATH}.random", "rng", "init_rng", seed=seed
@@ -297,7 +284,7 @@ def init_guassian_augmentation_flow() -> Flow:
             f"{_PLUGINS_IMPORT_PATH}.artifacts",
             "utils",
             "make_directories",
-            dirs=[def_data_dir],
+            dirs=[adv_data_dir],
         )
 
         log_mlflow_params_result = pyplugs.call_task(  # noqa: F841
@@ -311,35 +298,49 @@ def init_guassian_augmentation_flow() -> Flow:
             ),
         )
 
+        keras_classifier = pyplugs.call_task(
+            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.custom_fgm_plugins",
+            "registry_art",
+            "load_wrapped_tensorflow_keras_classifier",
+            name=model_name,
+            version=model_version,
+            imagenet_preprocessing=imagenet_preprocessing,
+            upstream_tasks=[init_tensorflow_results],
+        )
+
         distance_metrics_list = pyplugs.call_task(
             f"{_PLUGINS_IMPORT_PATH}.metrics",
             "distance",
             "get_distance_metric_list",
             request=DISTANCE_METRICS,
         )
+
         distance_metrics = pyplugs.call_task(
             f"{_CUSTOM_PLUGINS_IMPORT_PATH}.custom_fgm_plugins",
-            "defenses_image_preprocessing",
-            "create_defended_dataset",
-            def_type="gaussian_augmentation",
+            "attacks_fgm",
+            "create_adversarial_fgm_dataset",
             data_dir=testing_dir,
+            keras_classifier=keras_classifier,
             distance_metrics_list=distance_metrics_list,
-            def_data_dir=def_data_dir,
+            adv_data_dir=adv_data_dir,
             batch_size=batch_size,
             image_size=image_size,
-            augmentation=gaussian_augmentation_perform_data_augmentation,
-            ratio=gaussian_augmentation_ratio,
-            sigma=gaussian_augmentation_sigma,
-            apply_fit=gaussian_augmentation_apply_fit,
-            apply_predict=gaussian_augmentation_apply_predict,
+            rescale=rescale,
+            eps=eps,
+            eps_step=eps_step,
+            minimal=minimal,
+            norm=norm,
+            targeted=targeted,
+            target_index=target_index,
             upstream_tasks=[make_directories_results],
         )
+
         log_evasion_dataset_result = pyplugs.call_task(  # noqa: F841
             f"{_PLUGINS_IMPORT_PATH}.artifacts",
             "mlflow",
             "upload_directory_as_tarball_artifact",
-            source_dir=def_data_dir,
-            tarball_filename=def_tar_name,
+            source_dir=adv_data_dir,
+            tarball_filename=adv_tar_name,
             upstream_tasks=[distance_metrics],
         )
         log_distance_metrics_result = pyplugs.call_task(  # noqa: F841
@@ -365,4 +366,4 @@ if __name__ == "__main__":
     configure_structlog()
 
     with plugin_dirs(), StdoutLogStream(as_json), StderrLogStream(as_json):
-        _ = guassian_augmentation()
+        _ = fgm_attack()
