+++ examples/tensorflow-imagenet-resnet50/src/gaussian_augmentation.py
@@ -17,15 +17,13 @@
 # https://creativecommons.org/licenses/by/4.0/legalcode
 
 import os
-import tarfile
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, List
 
 import click
 import mlflow
 import numpy as np
 import structlog
-from mlflow.tracking import MlflowClient
 from prefect import Flow, Parameter
 from prefect.utilities.logging import get_logger as get_prefect_logger
 from structlog.stdlib import BoundLogger
@@ -74,13 +72,6 @@ def _coerce_int_to_bool(ctx, param, value):
 
 @click.command()
 @click.option(
-    "--data-dir",
-    type=click.Path(
-        exists=True, file_okay=False, dir_okay=True, resolve_path=True, readable=True
-    ),
-    help="Root directory for NFS mounted datasets (in container)",
-)
-@click.option(
     "--image-size",
     type=click.STRING,
     callback=_coerce_comma_separated_ints,
@@ -89,7 +80,7 @@ def _coerce_int_to_bool(ctx, param, value):
 @click.option(
     "--def-tar-name",
     type=click.STRING,
-    default="jpeg_compression_dataset.tar.gz",
+    default="guassian_augmentation_dataset.tar.gz",
     help="Name to give to tarfile artifact containing preprocessed  images",
 )
 @click.option(
@@ -105,34 +96,34 @@ def _coerce_int_to_bool(ctx, param, value):
     default=32,
 )
 @click.option(
-    "--jpeg-compression-quality",
-    type=click.INT,
-    help="The image quality, on a scale from 1 (worst) to 95 (best). Values above 95 should be avoided.",
-    default=50,
-)
-@click.option(
-    "--jpeg-compression-channels-first",
+    "--gaussian-augmentation-perform-data-augmentation",
     type=click.BOOL,
-    help="Set channels first or last.",
+    help="If set to true, original samples will be kept alongside noisy samples.",
     default=False,
 )
 @click.option(
-    "--jpeg-compression-apply-fit",
+    "--gaussian-augmentation-ratio",
+    type=click.FLOAT,
+    help="If data augmentation is set to true, specifies fraction of new samples (1=double dataset size).",
+    default=1,
+)
+@click.option(
+    "--gaussian-augmentation-sigma",
+    type=click.FLOAT,
+    help="Standard deviation of Gaussian noise to be added.",
+    default=1,
+)
+@click.option(
+    "--gaussian-augmentation-apply-fit",
     type=click.BOOL,
     help="Defense applied on images used for training.",
     default=False,
 )
 @click.option(
-    "--jpeg-compression-apply-predict",
+    "--gaussian-augmentation-apply-predict",
     type=click.BOOL,
     help="Defense applied on images used for testing.",
     default=True,
-)
-@click.option(
-    "--load-dataset-from-mlruns",
-    type=click.BOOL,
-    help="If set to true, instead loads the test dataset from a previous mlrun.",
-    default=False,
 )
 @click.option(
     "--dataset-run-id",
@@ -158,115 +149,108 @@ def _coerce_int_to_bool(ctx, param, value):
     help="Set the entry point rng seed",
     default=-1,
 )
-def jpeg_compression(
-    data_dir,
+def guassian_augmentation(
     image_size,
     def_tar_name,
     def_data_dir,
     batch_size,
-    jpeg_compression_quality,
-    jpeg_compression_channels_first,
-    jpeg_compression_apply_fit,
-    jpeg_compression_apply_predict,
-    load_dataset_from_mlruns,
+    gaussian_augmentation_perform_data_augmentation,
+    gaussian_augmentation_ratio,
+    gaussian_augmentation_sigma,
+    gaussian_augmentation_apply_fit,
+    gaussian_augmentation_apply_predict,
     dataset_run_id,
     dataset_tar_name,
     dataset_name,
     seed,
 ):
-
     LOGGER.info(
         "Execute MLFlow entry point",
-        entry_point="jpeg_compression",
-        data_dir=data_dir,
+        entry_point="guassian_augmentation",
         image_size=image_size,
         def_tar_name=def_tar_name,
         def_data_dir=def_data_dir,
         batch_size=batch_size,
-        jpeg_compression_quality=jpeg_compression_quality,
-        jpeg_compression_channels_first=jpeg_compression_channels_first,
-        jpeg_compression_apply_fit=jpeg_compression_apply_fit,
-        jpeg_compression_apply_predict=jpeg_compression_apply_predict,
-        load_dataset_from_mlruns=load_dataset_from_mlruns,
-        dataset_run_id=dataset_run_id,
-        dataset_tar_name=dataset_tar_name,
-        dataset_name=dataset_name,
+        gaussian_augmentation_perform_data_augmentation=gaussian_augmentation_perform_data_augmentation,
+        gaussian_augmentation_ratio=gaussian_augmentation_ratio,
+        gaussian_augmentation_sigma=gaussian_augmentation_sigma,
+        gaussian_augmentation_apply_fit=gaussian_augmentation_apply_fit,
+        gaussian_augmentation_apply_predict=gaussian_augmentation_apply_predict,
         seed=seed,
+        dataset_run_id=dataset_run_id,
+        dataset_name=dataset_name,
+        dataset_tar_name=dataset_tar_name,
     )
 
-    if load_dataset_from_mlruns:
-        data_dir = Path.cwd() / "dataset" / dataset_name
-        data_tar_name = dataset_tar_name
-        data_tar_path = download_image_archive(
-            run_id=dataset_run_id, archive_path=data_tar_name
-        )
-        with tarfile.open(data_tar_path, "r:gz") as f:
-            f.extractall(path=(Path.cwd() / "dataset"))
-
     with mlflow.start_run() as active_run:  # noqa: F841
-        flow: Flow = init_jpeg_compression_flow()
+        flow: Flow = init_guassian_augmentation_flow()
         state = flow.run(
             parameters=dict(
-                testing_dir=Path(data_dir),
                 image_size=image_size,
                 def_tar_name=def_tar_name,
                 def_data_dir=(Path.cwd() / def_data_dir).resolve(),
                 distance_metrics_filename="distance_metrics.csv",
                 batch_size=batch_size,
-                jpeg_compression_quality=jpeg_compression_quality,
-                jpeg_compression_channels_first=jpeg_compression_channels_first,
-                jpeg_compression_apply_fit=jpeg_compression_apply_fit,
-                jpeg_compression_apply_predict=jpeg_compression_apply_predict,
+                gaussian_augmentation_perform_data_augmentation=gaussian_augmentation_perform_data_augmentation,
+                gaussian_augmentation_ratio=gaussian_augmentation_ratio,
+                gaussian_augmentation_sigma=gaussian_augmentation_sigma,
+                gaussian_augmentation_apply_fit=gaussian_augmentation_apply_fit,
+                gaussian_augmentation_apply_predict=gaussian_augmentation_apply_predict,
                 seed=seed,
+                dataset_run_id=dataset_run_id,
+                dataset_name=dataset_name,
+                dataset_tar_name=dataset_tar_name,
             )
         )
 
     return state
 
 
-# Update data dir path if user is applying defense over image artifacts.
-def download_image_archive(
-    run_id: str, archive_path: str, destination_path: Optional[str] = None
-) -> str:
-    client: MlflowClient = MlflowClient()
-    image_archive_path: str = client.download_artifacts(
-        run_id=run_id, path=archive_path, dst_path=destination_path
-    )
-    LOGGER.info(
-        "Image archive downloaded",
-        run_id=run_id,
-        storage_path=archive_path,
-        dst_path=image_archive_path,
-    )
-    return image_archive_path
-
-
-def init_jpeg_compression_flow() -> Flow:
+def init_guassian_augmentation_flow() -> Flow:
     with Flow("Fast Gradient Method") as flow:
         (
-            testing_dir,
             image_size,
             def_tar_name,
             def_data_dir,
             distance_metrics_filename,
             batch_size,
-            jpeg_compression_quality,
-            jpeg_compression_channels_first,
-            jpeg_compression_apply_fit,
-            jpeg_compression_apply_predict,
+            gaussian_augmentation_perform_data_augmentation,
+            gaussian_augmentation_ratio,
+            gaussian_augmentation_sigma,
+            gaussian_augmentation_apply_fit,
+            gaussian_augmentation_apply_predict,
             seed,
+            dataset_run_id,
+            dataset_name,
+            dataset_tar_name,
         ) = (
-            Parameter("testing_dir"),
             Parameter("image_size"),
             Parameter("def_tar_name"),
             Parameter("def_data_dir"),
             Parameter("distance_metrics_filename"),
             Parameter("batch_size"),
-            Parameter("jpeg_compression_quality"),
-            Parameter("jpeg_compression_channels_first"),
-            Parameter("jpeg_compression_apply_fit"),
-            Parameter("jpeg_compression_apply_predict"),
+            Parameter("gaussian_augmentation_perform_data_augmentation"),
+            Parameter("gaussian_augmentation_ratio"),
+            Parameter("gaussian_augmentation_sigma"),
+            Parameter("gaussian_augmentation_apply_fit"),
+            Parameter("gaussian_augmentation_apply_predict"),
             Parameter("seed"),
+            Parameter("dataset_run_id"),
+            Parameter("dataset_name"),
+            Parameter("dataset_tar_name"),
+        )
+        data_tar_path = pyplugs.call_task(
+            f"{_PLUGINS_IMPORT_PATH}.artifacts",
+            "mlflow",
+            "download_all_artifacts_in_run",
+            run_id=dataset_run_id,
+            artifact_path=dataset_tar_name,
+        )
+        output_dir = pyplugs.call_task(
+            f"{_PLUGINS_IMPORT_PATH}.artifacts",
+            "utils",
+            "extract_tarfile_in_unique_subdir",
+            filepath=data_tar_path,
         )
         seed, rng = pyplugs.call_task(
             f"{_PLUGINS_IMPORT_PATH}.random", "rng", "init_rng", seed=seed
@@ -277,7 +261,7 @@ def init_jpeg_compression_flow() -> Flow:
         dataset_seed = pyplugs.call_task(
             f"{_PLUGINS_IMPORT_PATH}.random", "sample", "draw_random_integer", rng=rng
         )
-        init_tensorflow_results = pyplugs.call_task(
+        init_tensorflow_results = pyplugs.call_task(  # noqa: F841
             f"{_PLUGINS_IMPORT_PATH}.backend_configs",
             "tensorflow",
             "init_tensorflow",
@@ -311,16 +295,18 @@ def init_jpeg_compression_flow() -> Flow:
             f"{_CUSTOM_PLUGINS_IMPORT_PATH}.custom_fgm_plugins",
             "defenses_image_preprocessing",
             "create_defended_dataset",
-            def_type="jpeg_compression",
-            data_dir=testing_dir,
+            def_type="gaussian_augmentation",
+            data_dir=output_dir,
+            dataset_name=dataset_name,
             distance_metrics_list=distance_metrics_list,
             def_data_dir=def_data_dir,
             batch_size=batch_size,
             image_size=image_size,
-            quality=jpeg_compression_quality,
-            channels_first=jpeg_compression_channels_first,
-            apply_fit=jpeg_compression_apply_fit,
-            apply_predict=jpeg_compression_apply_predict,
+            augmentation=gaussian_augmentation_perform_data_augmentation,
+            ratio=gaussian_augmentation_ratio,
+            sigma=gaussian_augmentation_sigma,
+            apply_fit=gaussian_augmentation_apply_fit,
+            apply_predict=gaussian_augmentation_apply_predict,
             upstream_tasks=[make_directories_results],
         )
         log_evasion_dataset_result = pyplugs.call_task(  # noqa: F841
@@ -354,4 +340,4 @@ if __name__ == "__main__":
     configure_structlog()
 
     with plugin_dirs(), StdoutLogStream(as_json), StderrLogStream(as_json):
-        _ = jpeg_compression()
+        _ = guassian_augmentation()
