+++ examples/tensorflow-imagenet-resnet50/src/pt.py
@@ -15,17 +15,14 @@
 #
 # ACCESS THE FULL CC BY 4.0 LICENSE HERE:
 # https://creativecommons.org/licenses/by/4.0/legalcode
-
 import os
-import tarfile
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, List
 
 import click
 import mlflow
 import numpy as np
 import structlog
-from mlflow.tracking import MlflowClient
 from prefect import Flow, Parameter
 from prefect.utilities.logging import get_logger as get_prefect_logger
 from structlog.stdlib import BoundLogger
@@ -41,8 +38,9 @@ from dioptra.sdk.utilities.logging import (
     set_logging_level,
 )
 
-_CUSTOM_PLUGINS_IMPORT_PATH: str = "dioptra_custom"
 _PLUGINS_IMPORT_PATH: str = "dioptra_builtins"
+_CUSTOM_PLUGINS_IMPORT_PATH: str = "dioptra_custom"
+
 DISTANCE_METRICS: List[Dict[str, str]] = [
     {"name": "l_infinity_norm", "func": "l_inf_norm"},
     {"name": "l_1_norm", "func": "l_1_norm"},
@@ -87,16 +85,26 @@ def _coerce_int_to_bool(ctx, param, value):
     help="Dimensions for the input images",
 )
 @click.option(
-    "--def-tar-name",
+    "--adv-tar-name",
     type=click.STRING,
-    default="spatial_smoothing_dataset.tar.gz",
-    help="Name to give to tarfile artifact containing preprocessed  images",
+    default="testing_adversarial_pt.tar.gz",
+    help="Name to give to tarfile artifact containing pixel threshold images",
 )
 @click.option(
-    "--def-data-dir",
+    "--adv-data-dir",
     type=click.STRING,
     default="adv_testing",
-    help="Directory for saving preprocessed images",
+    help="Directory for saving pixel threshold images",
+)
+@click.option(
+    "--model-name",
+    type=click.STRING,
+    help="Name of model to load from registry",
+)
+@click.option(
+    "--model-version",
+    type=click.STRING,
+    help="Version of model to load from registry",
 )
 @click.option(
     "--batch-size",
@@ -105,46 +113,16 @@ def _coerce_int_to_bool(ctx, param, value):
     default=32,
 )
 @click.option(
-    "--spatial-smoothing-window-size",
+    "--th",
     type=click.INT,
-    help="The size of the sliding window for spatial smoothing defense.",
-    default=3,
+    help="Pixel Threshold Attack threshold value",
+    default=1,
 )
 @click.option(
-    "--spatial-smoothing-apply-fit",
-    type=click.BOOL,
-    help="Spatial smoothing applied on images used for training.",
-    default=False,
-)
-@click.option(
-    "--spatial-smoothing-apply-predict",
-    type=click.BOOL,
-    help="Spatial smoothing applied on images used for testing.",
-    default=True,
-)
-@click.option(
-    "--load-dataset-from-mlruns",
-    type=click.BOOL,
-    help="If set to true, instead loads the test dataset from a previous mlrun.",
-    default=False,
-)
-@click.option(
-    "--dataset-run-id",
-    type=click.STRING,
-    help="MLFlow Run ID of an updated dataset.",
-    default="",
-)
-@click.option(
-    "--dataset-tar-name",
-    type=click.STRING,
-    help="Name of dataset tarfile.",
-    default="adversarial_poison.tar.gz",
-)
-@click.option(
-    "--dataset-name",
-    type=click.STRING,
-    help="Name of dataset directory.",
-    default="adv_poison_data",
+    "--es",
+    type=click.INT,
+    help="Pixel Threshold Attack Evolution Strategy",
+    default=0,
 )
 @click.option(
     "--seed",
@@ -152,110 +130,91 @@ def _coerce_int_to_bool(ctx, param, value):
     help="Set the entry point rng seed",
     default=-1,
 )
-def spatial_smoothing(
+@click.option(
+    "--clip-values",
+    type=click.STRING,
+    callback=_coerce_comma_separated_ints,
+    help="Set the range for pixel values in the ART Estimator",
+)
+def pt_attack(
     data_dir,
     image_size,
-    def_tar_name,
-    def_data_dir,
+    adv_tar_name,
+    adv_data_dir,
+    model_name,
+    model_version,
     batch_size,
-    spatial_smoothing_window_size,
-    spatial_smoothing_apply_fit,
-    spatial_smoothing_apply_predict,
-    load_dataset_from_mlruns,
-    dataset_run_id,
-    dataset_tar_name,
-    dataset_name,
+    th,
+    es,
     seed,
+    clip_values,
 ):
-
     LOGGER.info(
         "Execute MLFlow entry point",
-        entry_point="spatial_smoothing",
+        entry_point="pt",
         data_dir=data_dir,
         image_size=image_size,
-        def_tar_name=def_tar_name,
-        def_data_dir=def_data_dir,
+        adv_tar_name=adv_tar_name,
+        adv_data_dir=adv_data_dir,
+        model_name=model_name,
+        model_version=model_version,
         batch_size=batch_size,
-        spatial_smoothing_window_size=spatial_smoothing_window_size,
-        spatial_smoothing_apply_fit=spatial_smoothing_apply_fit,
-        spatial_smoothing_apply_predict=spatial_smoothing_apply_predict,
-        load_dataset_from_mlruns=load_dataset_from_mlruns,
-        dataset_run_id=dataset_run_id,
-        dataset_tar_name=dataset_tar_name,
-        dataset_name=dataset_name,
+        th=th,
+        es=es,
         seed=seed,
+        clip_values=clip_values,
     )
 
-    if load_dataset_from_mlruns:
-        data_dir = Path.cwd() / "dataset" / dataset_name
-        data_tar_name = dataset_tar_name
-        data_tar_path = download_image_archive(
-            run_id=dataset_run_id, archive_path=data_tar_name
-        )
-        with tarfile.open(data_tar_path, "r:gz") as f:
-            f.extractall(path=(Path.cwd() / "dataset"))
-
     with mlflow.start_run() as active_run:  # noqa: F841
-        flow: Flow = init_spatial_smoothing_flow()
+        flow: Flow = init_pt_flow()
         state = flow.run(
             parameters=dict(
-                testing_dir=Path(data_dir),
+                testing_dir=Path(data_dir) / "testing",
                 image_size=image_size,
-                def_tar_name=def_tar_name,
-                def_data_dir=(Path.cwd() / def_data_dir).resolve(),
+                adv_tar_name=adv_tar_name,
+                adv_data_dir=(Path.cwd() / adv_data_dir).resolve(),
                 distance_metrics_filename="distance_metrics.csv",
+                model_name=model_name,
+                model_version=model_version,
                 batch_size=batch_size,
-                spatial_smoothing_window_size=spatial_smoothing_window_size,
-                spatial_smoothing_apply_fit=spatial_smoothing_apply_fit,
-                spatial_smoothing_apply_predict=spatial_smoothing_apply_predict,
+                th=th,
+                es=es,
                 seed=seed,
+                clip_values=clip_values,
             )
         )
 
     return state
 
 
-# Update data dir path if user is applying defense over image artifacts.
-def download_image_archive(
-    run_id: str, archive_path: str, destination_path: Optional[str] = None
-) -> str:
-    client: MlflowClient = MlflowClient()
-    image_archive_path: str = client.download_artifacts(
-        run_id=run_id, path=archive_path, dst_path=destination_path
-    )
-    LOGGER.info(
-        "Image archive downloaded",
-        run_id=run_id,
-        storage_path=archive_path,
-        dst_path=image_archive_path,
-    )
-    return image_archive_path
-
-
-def init_spatial_smoothing_flow() -> Flow:
-    with Flow("Fast Gradient Method") as flow:
+def init_pt_flow() -> Flow:
+    with Flow("Pixel Threshold") as flow:
         (
             testing_dir,
             image_size,
-            def_tar_name,
-            def_data_dir,
+            adv_tar_name,
+            adv_data_dir,
             distance_metrics_filename,
+            model_name,
+            model_version,
             batch_size,
-            spatial_smoothing_window_size,
-            spatial_smoothing_apply_fit,
-            spatial_smoothing_apply_predict,
+            th,
+            es,
             seed,
+            clip_values,
         ) = (
             Parameter("testing_dir"),
             Parameter("image_size"),
-            Parameter("def_tar_name"),
-            Parameter("def_data_dir"),
+            Parameter("adv_tar_name"),
+            Parameter("adv_data_dir"),
             Parameter("distance_metrics_filename"),
+            Parameter("model_name"),
+            Parameter("model_version"),
             Parameter("batch_size"),
-            Parameter("spatial_smoothing_window_size"),
-            Parameter("spatial_smoothing_apply_fit"),
-            Parameter("spatial_smoothing_apply_predict"),
+            Parameter("th"),
+            Parameter("es"),
             Parameter("seed"),
+            Parameter("clip_values"),
         )
         seed, rng = pyplugs.call_task(
             f"{_PLUGINS_IMPORT_PATH}.random", "rng", "init_rng", seed=seed
@@ -276,7 +235,7 @@ def init_spatial_smoothing_flow() -> Flow:
             f"{_PLUGINS_IMPORT_PATH}.artifacts",
             "utils",
             "make_directories",
-            dirs=[def_data_dir],
+            dirs=[adv_data_dir],
         )
 
         log_mlflow_params_result = pyplugs.call_task(  # noqa: F841
@@ -289,7 +248,15 @@ def init_spatial_smoothing_flow() -> Flow:
                 dataset_seed=dataset_seed,
             ),
         )
-
+        keras_classifier = pyplugs.call_task(
+            f"{_PLUGINS_IMPORT_PATH}.registry",
+            "art",
+            "load_wrapped_tensorflow_keras_classifier",
+            name=model_name,
+            version=model_version,
+            classifier_kwargs={"clip_values": clip_values},
+            upstream_tasks=[init_tensorflow_results],
+        )
         distance_metrics_list = pyplugs.call_task(
             f"{_PLUGINS_IMPORT_PATH}.metrics",
             "distance",
@@ -297,25 +264,26 @@ def init_spatial_smoothing_flow() -> Flow:
             request=DISTANCE_METRICS,
         )
         distance_metrics = pyplugs.call_task(
-            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.custom_fgm_plugins",
-            "defenses_image_preprocessing",
-            "create_defended_dataset",
+            f"{_CUSTOM_PLUGINS_IMPORT_PATH}.pixel_threshold",
+            "pixelthreshold",
+            "create_pt_dataset",
             data_dir=testing_dir,
+            keras_classifier=keras_classifier,
             distance_metrics_list=distance_metrics_list,
-            def_data_dir=def_data_dir,
+            adv_data_dir=adv_data_dir,
             batch_size=batch_size,
             image_size=image_size,
-            window_size=spatial_smoothing_window_size,
-            apply_fit=spatial_smoothing_apply_fit,
-            apply_predict=spatial_smoothing_apply_predict,
+            th=th,
+            es=es,
             upstream_tasks=[make_directories_results],
         )
+
         log_evasion_dataset_result = pyplugs.call_task(  # noqa: F841
             f"{_PLUGINS_IMPORT_PATH}.artifacts",
             "mlflow",
             "upload_directory_as_tarball_artifact",
-            source_dir=def_data_dir,
-            tarball_filename=def_tar_name,
+            source_dir=adv_data_dir,
+            tarball_filename=adv_tar_name,
             upstream_tasks=[distance_metrics],
         )
         log_distance_metrics_result = pyplugs.call_task(  # noqa: F841
@@ -341,4 +309,4 @@ if __name__ == "__main__":
     configure_structlog()
 
     with plugin_dirs(), StdoutLogStream(as_json), StderrLogStream(as_json):
-        _ = spatial_smoothing()
+        _ = pt_attack()
